requirements.txt
python-telegram-bot==5.3.0
pymysql==0.7.10
emoji==0.4.5
eloquent==0.5